+++ vingester-main.js
@@ -344,6 +344,7 @@ electron.app.on("ready", async () => {
         { iname: "C", itype: "string",  def: "2",           etype: "number",  ename: "Output2AudioChannels" },
         { iname: "o", itype: "string",  def: "0",           etype: "number",  ename: "Output2AudioDelay" },
         { iname: "n", itype: "boolean", def: true,          etype: "boolean", ename: "Output2SinkNDIEnabled" },
+        { iname: "v", itype: "boolean", def: true,          etype: "boolean", ename: "Output2SinkNDIAlpha" },
         { iname: "l", itype: "boolean", def: false,         etype: "boolean", ename: "Output2SinkNDITallyReload" },
         { iname: "m", itype: "boolean", def: false,         etype: "boolean", ename: "Output2SinkFFmpegEnabled" },
         { iname: "R", itype: "string",  def: "vbr",         etype: "string",  ename: "Output2SinkFFmpegMode" },